Refactor stats service
