+++ src/App/utils/testHelpers.ts
@@ -16,7 +16,7 @@ import PageViewsResolver from '../pageViews/pageViews.resolver'
 import PageViewsService from '../pageViews/pageViews.service'
 import RevalidatePageResolver from '../revalidatePage/revalidatePage.resolver'
 import { RevalidatePageService } from '../revalidatePage/revalidatePage.service'
-import StatsResolver from '../stats.resolver'
+import StatsResolver from '../WikiStats/wikiStats.resolver'
 import WikiSubscriptionResolver from '../subscriptions.resolver'
 import WikiSubscriptionService from '../subscriptions.service'
 import TagResolver from '../tag.resolver'
